pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ai-draw"
dynamic = ["version"]
description = "CLI tools for batch AI image generation via Gemini and OpenRouter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "Pillow>=10.0",
    "google-generativeai>=0.8",
    "requests>=2.31",
]

[project.scripts]
gemini-draw = "ai_draw.gemini:main"
or-draw = "ai_draw.openrouter:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "post-release"
local_scheme = "node-and-date"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_draw"]
